api_requirements.txt
# MillennialAi API Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Core MillennialAi dependencies (from existing requirements.txt)
torch>=2.0.0
transformers>=4.21.0
datasets>=2.14.0
accelerate>=0.21.0
peft>=0.5.0
bitsandbytes>=0.41.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
wandb>=0.15.0
tensorboard>=2.14.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
Pillow>=10.0.0
requests>=2.31.0
PyYAML>=6.0
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0

# Optional Azure dependencies (for cloud deployment)
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-storage-blob>=12.19.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0